invoice_reconciliation/requirements.txt
# Core Framework
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-groq>=1.0.0

# PDF Processing
pypdf>=4.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
pydantic>=2.0.0
rapidfuzz>=3.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0
